Extract geometry and cell creation to reservoir builder
