+++ ApplicationLibCode/ReservoirDataModel/CMakeLists_files.cmake
@@ -96,6 +96,7 @@ set(SOURCE_GROUP_HEADER_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RigWellAllocationOverTime.h
     ${CMAKE_CURRENT_LIST_DIR}/RigWellResultBranch.h
     ${CMAKE_CURRENT_LIST_DIR}/RigWellResultFrame.h
+    ${CMAKE_CURRENT_LIST_DIR}/RigReservoirBuilder.h
 )
 
 set(SOURCE_GROUP_SOURCE_FILES
@@ -190,6 +191,7 @@ set(SOURCE_GROUP_SOURCE_FILES
     ${CMAKE_CURRENT_LIST_DIR}/RigWellResultBranch.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RigWellResultFrame.cpp
     ${CMAKE_CURRENT_LIST_DIR}/RigDeclineCurveCalculator.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/RigReservoirBuilder.cpp
 )
 
 list(APPEND CODE_HEADER_FILES ${SOURCE_GROUP_HEADER_FILES})
